Move email authentification logic to /strategies/email
