+++ lib/habilitations/model.js
@@ -1,6 +1,6 @@
 const mongo = require('../util/mongo')
 const createError = require('http-errors')
-const {getCommuneEmail} = require('../util/communes')
+const {getCommuneEmail} = require('./strategies/email')
 
 async function createHabilitation({codeCommune, client}) {
   const now = new Date()
