+++ lib/habilitations/routes.js
@@ -6,10 +6,10 @@ const {communes} = require('../util/cog')
 const {ensureCodeCommune} = require('../util/middlewares')
 const w = require('../util/w')
 const {createAuthClient} = require('../clients')
-const {createHabilitation, askHabilitation, validateHabilitation, decreasesRemainingAttempts, rejectHabilitation, fetchHabilitation} = require('./model')
+const {createHabilitation, askHabilitation, validateHabilitation, fetchHabilitation} = require('./model')
 const {sendMail} = require('../util/sendmail')
 const createPinCodeEmail = require('../emails/pin-code')
-const {hasBeenSentRecently, generatePinCode, getExpirationDate} = require('../util/pin-code')
+const {pinCodeValidation, hasBeenSentRecently, generatePinCode, getExpirationDate} = require('./strategies/email')
 
 async function habilitationsRoutes(params = {}) {
   const app = new express.Router()
@@ -100,21 +100,10 @@ async function habilitationsRoutes(params = {}) {
       throw createError(403, 'Aucun code d’authentification n’est associé à cette habilitation')
     }
 
-    if (req.body.code !== strategy.pinCode) {
-      const {strategy} = await decreasesRemainingAttempts(req.habilitation._id)
+    const {validated, error} = await pinCodeValidation(req.body.code, req.habilitation)
 
-      if (strategy.remainingAttempts === 0) {
-        await rejectHabilitation()
-        return res.send({validated: false, error: 'Code non valide, demande d’habilitation rejetée'})
-      }
-
-      const plural = strategy.remainingAttempts > 1 ? 's' : ''
-      return res.send({validated: false, error: `Code non valide, ${strategy.remainingAttempts} tentative${plural} retestante${plural}`})
-    }
-
-    const now = new Date()
-    if (now > strategy.pinCodeExpiration) {
-      return res.send({validated: false, error: 'Code expiré'})
+    if (!validated) {
+      res.send({validated, error})
     }
 
     const habilitation = await validateHabilitation(req.habilitation)
